Add files via upload
GPA Calculator Program

This C# console application calculates the Grade Point Average (GPA) for a student based on their course credits and grades. The user is prompted to enter the student's name, credit hours, and grades for eight predefined courses. The program then computes the total credit hours and grade points, and subsequently the GPA. The results, including the course details and the final GPA, are displayed in a user-friendly format.

Features:
- Prompts for student's name, course credits, and grades.
- Supports five predefined courses: OOP, OOP Lab, DataBase, DataBase Lab, DLD, DLD Lab, Discrete, Expository.
- Accepts grades in the form of A (4) ,B (3), C (2), D (1).
- Calculates and displays the final GPA with two decimal precision.

**Usage:**
1. Run the program.
2. Follow the prompts to input the student's name, credit hours, and grades for each course.
3. View the calculated GPA and detailed course information.

This program is a simple and effective tool for students and educators to quickly calculate and understand GPA based on course performance.